+++ packages/back/src/Infrastructures/Adapters/Ogp/IOgpAdapter.ts
