+++ packages/back/src/Infrastructures/Adapters/Ogp/OgpAdapter.ts
